functions/generate-similarity-cache/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-aiplatform==1.*
numpy==1.*
vertexai==1.*
